refactor: move `staticStatusId` utility to `helper/document` module
